Sampler required DataSource object, eliminate FeatureDataset classes
