+++ examples/mistnetplus/mistnetplus_server.py
@@ -10,6 +10,7 @@ import torch
 os.environ['config_file'] = 'examples/mistnetplus/mistnet_lenet5_server.yml'
 
 from plato.config import Config
+from plato.datasources import feature
 from plato.samplers import all_inclusive
 from plato.servers import fedavg
 
@@ -48,7 +49,7 @@ class MistnetplusServer(fedavg.Server):
                 "[Server #%d] client #%d features received. Processing.",
                 os.getpid(), client_id)
             features = [self.client_payload[sid]]
-            feature_dataset = list(chain.from_iterable(features))
+            feature_dataset = feature.DataSource(features)
             sampler = all_inclusive.Sampler(feature_dataset)
             self.algorithm.train(feature_dataset, sampler,
                              Config().algorithm.cut_layer)
