+++ plato/algorithms/mistnet.py
@@ -18,19 +18,6 @@ from plato.utils import unary_encoding
 from plato.algorithms import fedavg
 
 
-class FeatureDataset(torch.utils.data.Dataset):
-    """Used to prepare a feature dataset for a DataLoader in PyTorch."""
-    def __init__(self, dataset):
-        self.dataset = dataset
-
-    def __len__(self):
-        return len(self.dataset)
-
-    def __getitem__(self, item):
-        image, label = self.dataset[item]
-        return image, label
-
-
 class Algorithm(fedavg.Algorithm):
     """The PyTorch-based MistNet algorithm, used by both the client and the
     server.
@@ -93,4 +80,4 @@ class Algorithm(fedavg.Algorithm):
 
     def train(self, trainset, sampler, cut_layer=None):
         """ Train the neural network model after the cut layer. """
-        self.trainer.train(FeatureDataset(trainset), sampler, cut_layer)
+        self.trainer.train(trainset.torch_dataset(), sampler, cut_layer)
